requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
binaryornot==0.4.4
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
cryptography==38.0.3
daphne==4.0.0
Django==4.1.3
django-browser-reload==1.6.0
django-environ==0.9.0
django-environ-2==2.3.0
django-model-utils==4.3.1
django-smtp-ssl==1.0
django-tailwind==3.4.0
django-widget-tweaks==1.4.12
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
redis==4.3.5
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
zope.interface==5.5.2
